requirements.txt
alembic==1.9.1
aniso8601==9.0.1
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.1.5
djangorestframework==3.14.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.1
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
mysqlclient==2.1.1
packaging==23.0
python-dotenv==0.21.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
tzdata==2022.7
Werkzeug==2.2.2
